requirements.txt
streamlit 
pypdf2 
fpdf  
streamlit-authenticator
pymupdf
google-cloud-aiplatform 



